requirements.txt
streamlit
yfinance
prophet
plotly
pandas
numpy
scikit-learn
torch
transformers
finnhub-python

langgraph
langchain
langchain-google-genai
langchain-community
duckduckgo-search
python-dotenv
streamlit-option-menu
streamlit-lottie
streamlit-navigation-bar
